+++ src/NoteGroup.cpp
@@ -47,12 +47,12 @@ float NoteGroup::get_staff_mid_point() const
 	return static_cast<float>(staff_position_sum) / static_cast<float>(m_notes.size());
 }
 
-float NoteGroup::get_staff_max() const
+int NoteGroup::get_staff_max() const
 {
 	return m_notes.back().get_staff_position();
 }
 
-float NoteGroup::get_staff_min() const
+int NoteGroup::get_staff_min() const
 {
 	return m_notes.front().get_staff_position();
 }
